Move M24C32 out of platform folder, have readWord, writeWord, etc use read/write bytes
